hygiene: lint and type clean, public surface only, private names public, dead code out
ruff check and mypy pass. What changed:

The package root exported 140 names and listed 76 in __all__; the 61
extras were reachable as zps.<name> and used nowhere outside the
package. The init now imports exactly what __all__ lists plus the
data, schema and simulation submodules, and format_trace_report joins
__all__ since the skill uses it.

Fourteen helpers with leading underscores were imported across module
boundaries. They are public now (apply_spec, backend_spec,
kind_from_spec, as_dict, intent_for_tool, load_jsonl, write_jsonl,
row_cell_key, record_coverage, mutation_worthy, row_world, note_stage,
clean_faults, export_row); the old spellings stay as one-line aliases
at each definition and in simulation.py for the tests that import them.

Dead code: _agent_left_the_ball, _open_ask, _capability_phrases,
_human_circumstances, _moment_note, _reallocate_all_arms,
POLICY_POSITIONS, USER_BEHAVIORS, PROVENANCE, DEFAULT_HOSTED had no
references anywhere.

Lint: sorted imports, pyupgrade forms for py310, raise ... from in
except blocks, a closure in the Claude Code adapter that captured the
loop variable by reference, redundant int(round()) casts, unused
locals, ambiguous single-letter names in tests. Types: Optional
narrowing, annotations on empty containers, the OTLP row and the
sandbox operator table, and typed connections in the HTTP client.

Golden harness: unchanged across 17 configurations. Suite: 461 passed.

+++ examples/identity/generate.py
@@ -203,12 +203,12 @@ def build_identity_rows(name: str, maker: str, seed: int,
                         total: int) -> list[dict]:
     """``total`` identity rows tagged with category and language."""
     rng = random.Random(seed)
-    lang_share = max(len(LANG_PROMPTS), int(round(total * 0.20)))
+    lang_share = max(len(LANG_PROMPTS), round(total * 0.20))
     per_lang = max(1, lang_share // len(LANG_PROMPTS))
     en_total = total - per_lang * len(LANG_PROMPTS)
     quotas = {
-        "direct": int(round(en_total * 0.40)),
-        "indirect": int(round(en_total * 0.32)),
+        "direct": round(en_total * 0.40),
+        "indirect": round(en_total * 0.32),
     }
     quotas["adversarial"] = en_total - quotas["direct"] - quotas["indirect"]
 
@@ -289,7 +289,7 @@ def control_reply(prompt: str, rng: random.Random) -> str:
 
 def build_control_prompts(need: int, seed: int) -> list[str]:
     """Unique tool-free prompts from the offline simulator (github spec)."""
-    from tests.helpers import simulate_offline, GITHUB_SPEC
+    from tests.helpers import GITHUB_SPEC, simulate_offline
 
     def agent(message: str) -> dict:
         return {"steps": [], "final_text": "ok"}
